+++ CHANGES.md
@@ -9,3 +9,4 @@
 - **Video Player:** Added volume contorl
 - **Exchange Calculator:** UI change, API change
 - **Landing Page with Navbar and Modal:** UI change, modal fading out added
+- **Hangman Game:** Keyboard and hints added